Use relative imports everywhere
When installing the plugin from a ZIP file, the directory name containing
the plugin might be different that the expected one (i.e. `estimap_recreation`)
and as a result, the absolute import was failing like this: https://i.imgur.com/eCmSXzj.png

By using a relative import, the import should work, no matter how the
directory is named.